+++ NEWS.md
@@ -7,11 +7,13 @@
 ## lidR v2.0.1 (Release date: 2010-02-02)
 
 - Change: the function `catalog` has been slightly modified in prevision of the release of the package `rlas 1.3.0` to preserve future compatibility. This is invisible for the users.
+- New: `lasnormalize` gained a parameter `na.rm = TRUE`
 - Fix: an error occurend when plotting a LAScatalog with the option `chunk_pattern = TRUE`: objet 'ctg' introuvable.
 - Fix: examples in documentation of `tin()` and `knnidw()` were inverted.
 - Fix: [#213](https://github.com/Jean-Romain/lidR/issues/213) bug when using option `keep_lowest` in `grid_terrain`.
 - Fix: [#212](https://github.com/Jean-Romain/lidR/issues/212) bug when merging big rasters that exeed the memory allowed by the raster package
 - Fix: bug when merging rasters when some of then only have one cell
+- Fix: bug when printing a 0 point LAS object
 
 # lidR v2.0.0 (Release date: 2019-01-02)